requirements.txt
pydantic~=2.9.2
openai~=1.51.2
anthropic~=0.36.0
aiosqlite~=0.20.0
llama-cpp-python~=0.3.1
python-dotenv~=1.0.1
fastapi~=0.115.0
starlette~=0.38.6
uvicorn~=0.31.1
pytest-asyncio~=0.24.0
